Basic Sheep Movement
The player is now able to move the sheep once it is spawned BUT it is far from finished. I simply took elements from the player movement and copy and pasted the elements I wanted for my sheep.

Things that need to be done in order for sheep spawning/movement to be fully complete:
-Only be able to move sheep if player has left clicked on it
-Disable movement of all other entities that are not selected
-Adjust values so as to make sheep movement feel satisfying
-Only allow player to spawn 3 sheep total and change spawn buttons to 1, 2 , and 3 (each sheep will have a different "ability")