Sheep Sprite
Imported sheep sprite with an idle and walk animation. I attached it to the prototype "square" prefab I had as a sheep placeholder and now I am getting an error. I'm not sure why this is at all because I didn't really touch anything besides that but hopefully it's an easy fix