Sheep Update
Fixed the spawning system, it now handles all possible errors (at least from what I've tested) and is done with the mouse wheel which feels a lot better. I finally figured out the snap mechanic for the sheep and it's feeling pretty decent. I added a little timer so that the snap actually has time to settle in without any miss user input. Finally, I made it so that when a sheep is on top of another sheep, the idle animation will play and not the walking one (might have to fix this in the future so that the sheep start playing the walk animation as soon as they are no longer "snapped' but I'll think about that later)